finished part4 	modified:   app.js 	modified:   controllers/blogs.js 	new file:   controllers/login.js 	new file:   controllers/users.js 	modified:   models/blog.js 	new file:   models/user.js 	modified:   package-lock.json 	modified:   package.json 	new file:   requests/blogs/create_blog.rest 	renamed:    requests/delete_blog.rest -> requests/blogs/delete_blog.rest 	renamed:    requests/get_blogs.rest -> requests/blogs/get_blogs.rest 	deleted:    requests/create_blog.rest 	new file:   requests/login/login.rest 	new file:   requests/users/create_user.rest 	modified:   tests/bloglist_api.test.js 	modified:   tests/test_helper.js 	new file:   tests/users_api.test.js 	modified:   utils/middleware.js

+++ controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({})
+    .populate('user', { username: 1, name: 1 })
   res.json(blogs)
 })
 
@@ -15,23 +17,37 @@ blogsRouter.get('/:id', async (req, res) => {
   }
 })
 
-blogsRouter.post('/', async (req, res) => {
-  const body = req.body
+blogsRouter.post('/', userExtractor, async (req, res) => {
+  const { title, author, url, likes } = req.body
+  const user = req.user
 
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes
+    title,
+    author,
+    url,
+    likes: likes || 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
   res.status(201).json(savedBlog)
-
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', userExtractor, async (req, res) => {
+  const user = req.user
+  const blog = await Blog.findById(req.params.id)
+
+  if (!blog) {
+    return res.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return res.status(401).json({ error: 'only the creator can delete this blog' })
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
